refactor(demo): withserver broken please help
